Add Beats in Bar both ways
Each tatum knows what beat it is contained in. Each beat knows the
tatums inside it, and the bar it is contained in. Every bar knows the
beats inside it and the section it is contained in. Each section knows
what bars it contains. For tatums the array list contained is just
empty. For Sections the containedIn int is -1